backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
python-docx
openai
langchain
langchain-text-splitters
langchain_community
sentence-transformers
faiss-cpu
numpy